medical_bot_fullstack/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
google-generativeai
python-dotenv
pydantic
psycopg2-binary # Optional for PostgreSQL
